feat: add link transformation functions for Twitter, Instagram, and TikTok; implement reaction handling for various keywords
