+++ package.json
@@ -6,6 +6,7 @@
   "main": "index.js",
   "scripts": {
     "build": "tsc",
+    "start": "tsc && node build/index.js",
     "dev": "nodemon --watch src --ext ts --exec \"node --loader ts-node/esm/transpile-only src/index.ts\""
   },
   "repository": {
